(tistory @2341) getCategoryId 이름이 escaping 되는 경우 들고 오지 못하는 버그 수정 addCategory 부모 엄하게 찾는 버그 수정 deleteCategory 파라미터 확인 강화 modifyCategory 이스케이핑 되는 이름의 경우 중복 가능한 버그 수정

+++ lib/model/category.php
@@ -2,10 +2,14 @@
 
 function getCategoryId($owner, $name, $parentName = false) {
 	global $database;
+	
+	$name = mysql_real_escape_string($name);
 	if ($parentName === false)
 		$sql = "SELECT id FROM {$database['prefix']}Categories WHERE owner = $owner AND name = '$name'";
-	else
+	else {
+		$parentName = mysql_real_escape_string($parentName);
 		$sql = "SELECT c.id FROM {$database['prefix']}Categories c LEFT JOIN {$database['prefix']}Categories c2 ON c.parent = c2.id AND c.owner = c2.owner WHERE c.owner = $owner AND c.name = '$name' AND c2.name = '$parentName'";
+	}
 	return fetchQueryCell($sql);
 }
 
@@ -79,8 +83,10 @@ function addCategory($owner, $parent, $name) {
 	global $database;
 	if (empty($name))
 		return false;
+	if (!is_numeric($parent))
+		return false;
 	if ($parent !== null) {
-		$label = fetchQueryCell("SELECT name FROM {$database['prefix']}Categories WHERE owner = $owner AND parent IS NULL");
+		$label = fetchQueryCell("SELECT name FROM {$database['prefix']}Categories WHERE owner = $owner AND parent = $parent");
 		if ($label === null)
 			return false;
 		$label .= '/' . $name;
@@ -111,6 +117,9 @@ function addCategory($owner, $parent, $name) {
 
 function deleteCategory($owner, $id) {
 	global $database;
+	
+	if (!is_numeric($id))
+		return false;
 	return executeQuery("DELETE FROM {$database['prefix']}Categories WHERE owner = $owner AND id = $id");
 }
 
@@ -124,16 +133,15 @@ function modifyCategory($owner, $id, $name) {
 	$label = $row['name'];
 	$parentId = $row['id'];	
 	if (!empty($parentId)) {
-		$parentStr = "AND parent=$parentId";
+		$parentStr = "AND parent = $parentId";
 	} else
 		$parentStr = 'AND parent is null';
 	
+	$label = mysql_real_escape_string(empty($label) ? $name : "$label/$name");
+	$name = mysql_real_escape_string($name);
 	$sql = "SELECT count(*) FROM {$database['prefix']}Categories WHERE owner = $owner AND name='$name' $parentStr";	
 	if(fetchQueryCell($sql) >0)
 		return false;	
-		
-	$label = mysql_escape_string(empty($label) ? $name : "$label/$name");
-	$name = mysql_escape_string($name);
 	
 	$result = mysql_query("UPDATE {$database['prefix']}Categories SET name = '$name', label = '$label' WHERE owner = $owner AND id = $id");
 	if ($result && (mysql_affected_rows() > 0))
